#1 Add readiness, liveness
