+++ xtermserver/xtermserver.go
@@ -62,7 +62,7 @@ func (httpServer *XtermServerImpl) Serve(ctx context.Context) error {
 		PathMetrics:          httpServer.PathMetrics,
 		PathReadiness:        httpServer.PathReadiness,
 		PathXtermjs:          httpServer.PathXtermjs,
-		ServerAddr:           httpServer.ServerAddr,
+		ServerAddress:        httpServer.ServerAddr,
 		ServerPort:           httpServer.Port,
 		WorkingDir:           httpServer.WorkingDir,
 	}
